move common form input component to common/
